Split every repo-policy-check handler to its own file
I'll probably move to grouping these in fewer files.
